Getting rid of InputColumnsManager and placing input column fetch logic within QueryManager
