+++ mathesar_ui/src/systems/query-builder/column-selection-pane/SelectableColumn.svelte
@@ -2,7 +2,7 @@
   import { createEventDispatcher } from 'svelte';
   import { Button } from '@mathesar-component-library';
   import ColumnName from '@mathesar/components/column/ColumnName.svelte';
-  import type { ColumnWithLink } from '../InputColumnsManager';
+  import type { ColumnWithLink } from '../QueryManager';
 
   const dispatch = createEventDispatcher();
